backend/requirements.txt
Django==3.1.1
djangorestframework
markdown
pygments
django-filter
coreapi
requests==2.25.1
django-cors-headers
